% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchmsqrob2PTMGUI.R
\name{launchmsqrob2PTMGUI}
\alias{launchmsqrob2PTMGUI}
\title{Launches the msqrob2PTMGUI Shiny App}
\usage{
launchmsqrob2PTMGUI(maxSize = 500)
}
\arguments{
\item{maxSize}{maximum memory size that input files are allowed to have in Mb}
}
\value{
shiny application object
}
\description{
Launches the msqrob2PTMGUI Shiny App
}
